+++ src/main/java/com/dnd8th/dao/review/ReviewFindDao.java
@@ -1,4 +1,4 @@
-package com.dnd8th.dao;
+package com.dnd8th.dao.review;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
